modify download code

+++ data/download.py
@@ -4,7 +4,7 @@ import os
 
 
 # default setting would download all trajacories to /data/tartanairv2filtered/
-destination_folder_base = 'tartanairv2filtered'
+destination_folder_base = '/mnt/data/tartanairv2filtered'
 
 # Define your Azure Blob Storage connection string and container name
 connection_string = 'DefaultEndpointsProtocol=https;AccountName=tartanairv2;AccountKey=PH1q1TB4fHqGFnvfivj9jPrvABz2ESX1OKkrA4+8G3qoHJpIPDmDok4d2uzDNF7RVR9A4cE7Y00n5nsCrc+edA==;EndpointSuffix=core.windows.net'
@@ -23,7 +23,7 @@ def download():
 
     # path_combinations = ['depth_lcam_fish', 'depth_rcam_fish', 'image_lcam_fish',
     #                      'image_rcam_fish', 'imu', 'lidar', 'pose_rcam_front.txt', 'pose_lcam_front.txt']
-    path_combinations = ['depth_lcam_fish', 'image_lcam_fish', 'imu', 'lidar', 'pose_lcam_front.txt']
+    path_combinations = ['depth_lcam_fish', 'image_lcam_fish', 'lidar', 'pose_lcam_front.txt']
 
     blob_list = blob_container_client.list_blobs()
 
@@ -33,7 +33,6 @@ def download():
             print(blob.name) #e.g., AbandonedCableExposure/Data_easy/P000/depth_lcam_fish/000860_lcam_fish_depth.png
 
             blob_client = blob_container_client.get_blob_client(blob.name)
-            blob_data = blob_client.download_blob()
 
             filename = blob.name.split("/")[-1]
             trajectory_folder = os.path.join(destination_folder_base, f'{path_lists[0]}/{path_lists[1]}/{path_lists[2]}')
@@ -48,9 +47,11 @@ def download():
             # Create the parent directory (and any intermediate directories) if it doesn't exist
             os.makedirs(destination_folder, exist_ok=True)
 
-            print(f"destination_file: {destination_file}")
-            with open(destination_file, "wb") as file:
-                file.write(blob_data.readall())
+            if not os.path.exists(destination_file):
+                print(f"destination_file: {destination_file}")
+                blob_data = blob_client.download_blob()
+                with open(destination_file, "wb") as file:
+                    file.write(blob_data.readall())
 
 
 if __name__ == '__main__':
